LayoutTests/css3/filters/filter-property-parsing-expected.txt

Test the parsing of the -webkit-filter property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Basic reference : url("#a")
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "url(\"#a\")"

Bare unquoted reference converting to quoted form : url(#a)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "url(\"#a\")"

Multiple references : url('#a') url('#b')
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "url(\"#a\") url(\"#b\")"

Reference as 2nd value : grayscale(1) url("#a")
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "grayscale(1) url(\"#a\")"

Integer value : grayscale(1)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "grayscale(1)"

Percentage value : grayscale(50%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "grayscale(50%)"

Float value converts to integer : grayscale(1.0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "grayscale(1)"

Zero value : grayscale(0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "grayscale(0)"

No values : grayscale()
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "grayscale()"

Multiple values : grayscale(0.5) grayscale(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "grayscale(0.5) grayscale(0.25)"

Integer value : sepia(1)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "sepia(1)"

Percentage value : sepia(50%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "sepia(50%)"

Float value converts to integer : sepia(1.0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "sepia(1)"

Zero value : sepia(0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "sepia(0)"

No values : sepia()
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "sepia()"

Multiple values : sepia(0.5) sepia(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "sepia(0.5) sepia(0.25)"

Rule combinations : sepia(0.5) grayscale(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "sepia(0.5) grayscale(0.25)"

Integer value : saturate(1)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "saturate(1)"

Percentage value : saturate(50%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "saturate(50%)"

Percentage value > 1 : saturate(250%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "saturate(250%)"

Float value converts to integer : saturate(1.0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "saturate(1)"

Input value > 1 : saturate(5.5)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "saturate(5.5)"

Zero value : saturate(0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "saturate(0)"

No values : saturate()
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "saturate()"

Multiple values : saturate(0.5) saturate(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "saturate(0.5) saturate(0.25)"

Rule combinations : saturate(0.5) grayscale(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "saturate(0.5) grayscale(0.25)"

Degrees value as integer : hue-rotate(10deg)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "hue-rotate(10deg)"

Degrees float value converts to integer : hue-rotate(10.0deg)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "hue-rotate(10deg)"

Radians value : hue-rotate(10rad)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "hue-rotate(10rad)"

Gradians value : hue-rotate(10grad)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "hue-rotate(10grad)"

Turns value : hue-rotate(0.5turn)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "hue-rotate(0.5turn)"

Zero value : hue-rotate(0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "hue-rotate(0deg)"

No values : hue-rotate()
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "hue-rotate()"

Rule combinations : hue-rotate(10deg) grayscale(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "hue-rotate(10deg) grayscale(0.25)"

Integer value : invert(1)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "invert(1)"

Percentage value : invert(50%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "invert(50%)"

Float value converts to integer : invert(1.0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "invert(1)"

Zero value : invert(0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "invert(0)"

No values : invert()
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "invert()"

Multiple values : invert(0.5) invert(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "invert(0.5) invert(0.25)"

Rule combinations : invert(0.5) grayscale(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "invert(0.5) grayscale(0.25)"

Integer value : opacity(1)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "opacity(1)"

Percentage value : opacity(50%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "opacity(50%)"

Float value converts to integer : opacity(1.0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "opacity(1)"

Zero value : opacity(0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "opacity(0)"

No values : opacity()
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "opacity()"

Multiple values : opacity(0.5) opacity(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "opacity(0.5) opacity(0.25)"

Rule combinations : opacity(0.5) grayscale(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "opacity(0.5) grayscale(0.25)"

Integer value : brightness(1)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "brightness(1)"

Percentage value : brightness(50%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "brightness(50%)"

Float value converts to integer : brightness(1.0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "brightness(1)"

Zero value : brightness(0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "brightness(0)"

No values : brightness()
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "brightness()"

Multiple values : brightness(0.5) brightness(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "brightness(0.5) brightness(0.25)"

Rule combinations : brightness(0.5) grayscale(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "brightness(0.5) grayscale(0.25)"

Parameter more than 100% : brightness(101%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "brightness(101%)"

Rule combinations : grayscale(0.25) brightness(0.5)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "grayscale(0.25) brightness(0.5)"

Integer value : contrast(1)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "contrast(1)"

Percentage value : contrast(50%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "contrast(50%)"

Percentage value > 1 : contrast(250%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "contrast(250%)"

Float value converts to integer : contrast(1.0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "contrast(1)"

Zero value : contrast(0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "contrast(0)"

No values : contrast()
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "contrast()"

Value greater than one : contrast(2)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "contrast(2)"

Multiple values : contrast(0.5) contrast(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "contrast(0.5) contrast(0.25)"

Rule combinations : contrast(0.5) grayscale(0.25)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "contrast(0.5) grayscale(0.25)"

Rule combinations : grayscale(0.25) contrast(0.5)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "grayscale(0.25) contrast(0.5)"

One zero to px : blur(0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "blur(0px)"

One length : blur(10px)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "blur(10px)"

No values : blur()
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "blur()"

Color then three values : drop-shadow(red 1px 2px 3px)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "drop-shadow(red 1px 2px 3px)"

Three values then color : drop-shadow(1px 2px 3px red)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "drop-shadow(red 1px 2px 3px)"

Color then three values with zero length : drop-shadow(#abc 0 0 0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "drop-shadow(rgb(170, 187, 204) 0px 0px 0px)"

Three values with zero length : drop-shadow(0 0 0)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "drop-shadow(0px 0px 0px)"

Two values no color : drop-shadow(1px 2px)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "drop-shadow(1px 2px)"

Multiple operations : grayscale(0.5) sepia(0.25) saturate(0.75) hue-rotate(35deg) invert(0.2) opacity(0.9) blur(5px) drop-shadow(green 1px 2px 3px)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is "grayscale(0.5) sepia(0.25) saturate(0.75) hue-rotate(35deg) invert(0.2) opacity(0.9) blur(5px) drop-shadow(green 1px 2px 3px)"
PASS successfullyParsed is true

TEST COMPLETE
